d_total_weight_from_packaging: split common test class
